Cirrus-CI: temporarily comment out Debian task
Debian includes a newer GCC which warns about a number of instances of
strncpy in ELF Tool Chain.  All appear to be false positives; some are
cases of an overly aggressive warning (and the code should not change)
while in other cases a change to address the warning would also improve
readability.  In any case for now just remove the Debian task from
Cirrus-CI as it is not an officially supported platform.